requirements.txt
PyYAML==6.0.1
requests==2.31.0Markdown==3.5.1
